Add files via upload
The project objective is to predict whether a customer will purchase a car based on various factors such as salary, age, and other independent variables. To achieve this, the data from 'Social_Network_Ads.csv' is used. Logistic Regression, a widely used classification algorithm, is employed in this scenario. Since the output variable is binary (0 or 1, representing not buying and buying the car, respectively), Logistic Regression is suitable for this predictive task.

By training the model on historical data with known purchase outcomes, the marketing team can use the predictions to target potential customers more effectively. By identifying customers with a higher likelihood of making a purchase, targeted marketing strategies can be implemented to increase sales and optimize marketing efforts. This data-driven approach enables companies or show-room managers to make informed decisions and improve overall sales performance.